backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.0
google-genai>=1.11.0
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.0
redis==5.0.8
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
pillow==10.4.0
httpx>=0.28.1
numpy>=1.26.4
python-multipart==0.0.12
